+++ Hernandez - Menu/Assets/Scripts/LightPanel.cs
@@ -1,40 +1,18 @@
-
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
 public class LightPanel : MonoBehaviour
 {
-    public int verifica;
-    private bool taCerto = false;
-    [SerializeField] private theLightSwitch lightSwitch_1;
-    [SerializeField] private theLightSwitch lightSwitch_2;
-    [SerializeField] private theLightSwitch lightSwitch_3;
-    [SerializeField] private theLightSwitch lightSwitch_4;
-    [SerializeField] private theLightSwitch lightSwitch_5;
-    
-    void Update()
+    public int verifica = 0;
+    // Start is called before the first frame update
+    void Start()
     {
-        VerificaLigar();
+        
     }
 
-    public bool VerificaLigar()
+    // Update is called once per frame
+    void Update()
     {
-        if ((lightSwitch_1.isOn == true) && (lightSwitch_2.isOn == true) && (lightSwitch_3.isOn == true) && (lightSwitch_4.isOn == true) && (lightSwitch_5.isOn == true))
-        {
-            if(!taCerto){
-                taCerto = true;
-                this.verifica++;
-                Debug.Log(verifica);
-            }
-            return true;
-        }
-        //Caso os switchs não estejam mais ligados :(
-        if(taCerto){
-            taCerto = false;
-            this.verifica--;
-            Debug.Log(verifica);
-        }
-        return false;
     }
 }
